[refactor] base.AvPacket的时间戳类型修改为int64, avpacket_stream.go移入package base

+++ pkg/base/avpacket.go
@@ -44,7 +44,7 @@ func (a AvPacketPt) ReadableString() string {
 //
 type AvPacket struct {
 	PayloadType AvPacketPt
-	Timestamp   uint32 // TODO(chef): 改成int64
+	Timestamp   int64
 	Payload     []byte
 }
